vi: fix starting vi without filename
Since commit 32afd3a (vi: some simplifications) starting vi without a
filename on the command line has resulted in an error message in the
status line:  '(null)' Bad address.

The commit removed a call to the file_size function which checked for a
null filename.  This prevented file_insert (and hence open) from being
called with a null filename.

Signed-off-by: Ron Yorston <rmy@tigress.co.uk>