Revert "modprobe,rmmod: reject module names with slashes"
This reverts commit 4e314faa0aecb66717418e9a47a4451aec59262b.

Reported-by: Natanael Copa <ncopa@alpinelinux.org>
The problem with this is that there are some module aliases from
kernel that contains slashes. dmi modules to be more exact. Since
Alpine linux searches for 'modalias' in /sys and calls modprobe -a with
all found aliases, it would stop loading modules when it came to dmi
modalias with slash. This resulted in disk controller not beeing loaded.

http://thread.gmane.org/gmane.linux.distributions.alpine.devel/2790/focus=2793